mid day progress
