--- ucsb_drawings/drawsearch/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class DrawsearchConfig(AppConfig):
-    name = 'drawsearch'
