--- ucsb_drawings/drawsearch/templates/drawsearch/search_form.html
@@ -1,9 +1,0 @@
-<!-- https://developer.mozilla.org/en-US/docs/Learn/Server-side/Django/Forms -->
-
-
-<form action="/team_name_url/" method="post">
-    <label for="team_name">Enter name: </label>
-    <input id="team_name" type="text" name="name_field" value="Default name for team.">
-    <input type="submit" value="OK">
-</form>
-
