--- ucsb_drawings/drawsearch/urls.py
@@ -1,13 +1,0 @@
-from django.urls import path
-
-from . import views
-
-urlpatterns = [
-    path('', views.index, name='search'),
-
-
-]
-
-
-
-
