--- ucsb_drawings/drawsearch/views.py
@@ -1,9 +1,0 @@
-from django.http import HttpResponse
-from django.shortcuts import render
-
-    
-
-    
-def index(request):
-    return HttpResponse("Hello, world. You're at the Search index.")
-
