--- ucsb_drawings/ucsb_drawings/wsgi.py
@@ -1,16 +1,0 @@
-"""
-WSGI config for ucsb_drawings project.
-
-It exposes the WSGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/2.1/howto/deployment/wsgi/
-"""
-
-import os
-
-from django.core.wsgi import get_wsgi_application
-
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'ucsb_drawings.settings')
-
-application = get_wsgi_application()
